+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change log
 This is the changelog for [AWS Architect](readme.md).
 
+## 2.1 ##
+* Add support for creating dynamoDB Tables.
+
 ## 2.0 ##
 * All configuration is now contained in the api files. Generation, run, and testing is completed via index.js.  New templates created.
 * A single lambda function is created, for the purpose of the whole microservice.